rename: 패키지 model -> domain으로 변경
- 단순히 모델 객체만 포함시키지 않고 repository등 다양한 객체들을 포함하기 위해 변경
