+++ src/main/java/com/wootecam/jspcafe/domain/Question.java
@@ -1,4 +1,4 @@
-package com.wootecam.jspcafe.model;
+package com.wootecam.jspcafe.domain;
 
 import java.time.LocalDateTime;
 import java.util.Objects;
