+++ src/main/java/com/wootecam/jspcafe/service/QuestionService.java
@@ -1,6 +1,6 @@
 package com.wootecam.jspcafe.service;
 
-import com.wootecam.jspcafe.model.Question;
+import com.wootecam.jspcafe.domain.Question;
 import com.wootecam.jspcafe.repository.QuestionRepository;
 import java.time.LocalDateTime;
 import java.util.List;
