+++ src/main/java/com/wootecam/jspcafe/servlet/question/QuestionDetailServlet.java
@@ -1,6 +1,6 @@
 package com.wootecam.jspcafe.servlet.question;
 
-import com.wootecam.jspcafe.model.Question;
+import com.wootecam.jspcafe.domain.Question;
 import com.wootecam.jspcafe.service.QuestionService;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
